[INJIMOB-3409] refactor: extract constants to file
