+++ Sources/OpenID4VP/Constants/DIDMethod.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum DIDMethod: String {
+    case web = "web"
+    case key = "key"
+    case jwk = "jwk"
+}
