+++ Sources/OpenID4VP/Constants/PublicKeyType.swift
@@ -1,0 +1,7 @@
+import Foundation
+import CryptoKit
+
+enum PublicKeyType {
+    case secKey(SecKey)
+    case ed25519(Curve25519.Signing.PublicKey)
+}
