+++ Tests/OpenID4VPTests/jwt/KeyResolver/DidPublicKeyResolverTests.swift
@@ -240,7 +240,4 @@ class DidPublicKeyResolverTests: XCTestCase {
             )
         }
     }
-
-
-
 }
